requirements.txt
Django==2.2.7
django-filter==2.2.0
django-heroku==0.3.1
django-extensions==2.2.9
djangorestframework==3.10.3
django-cors-headers==3.4.0
Pillow==7.1.2
